Added pointer dereference function for structs and blobs
Added C_Ptr helper struct. This is useful when a function recieves a pointer and updates it (instead of return a value), and we want to extract the updated value

Added example of loading a raw image

Added example of loading PNG image and converting it to raw image. This requires dereferencing the result to swap R and B on every pixel