Allow accessing user_data from micropython side
Added 'cast' function to Blob. Optional parameter is dest type. Casting user_data will return either None or dict

Allow converting micropython object to pointer. This is needed for setting a new dict as user_data which is a void ptr.

Bugfix: When conveting dict to struct, raise an exception when field is not found