Add support for exporting int constants
An int constant is defined as an enum with the name "ENUM_LV_...". It will be handled as an enum and grouped according to prefix, unless its unique and in such case will be placed as an integer on the module itself.

Also update lvgl to dev-6.1 and generate lv_mpy_example.c. An update to lv_conf.h was also needed for dev-6.1