fix(ili9341): fix invert kwarg
Prior to this commit, the `ili9341` class `__init__` method would raise
a `KeyError` exception if `invert==True`. The issue was that the
following conditional did not add the required `"data"` key/value pair
to the `init_cmds` list item:

    if invert:
        self.init_cmds.append({'cmd': 0x21})

In this commit, add the required `"data"` key/value pair to the
`init_cmds` list item when `invert==True`.