gen_mpy.py: Convert a callback in a funcptr back to Python function, if possible
When setting a function pointer (such as 'flush_cb') to a python callback it can be used in C as a function pointer. But when reading it again in Micropython we would like to get the original Python function and not the callback wrapper. This is important becuase the callback wrapper is relying on 'user_data' that might change or not be available, but the original Python function is independant

Update gen/lv_mpy_example.c, but remove gen/lv_mpy_example.json since it's not really used or useful for anything right now (although the script can still generate it)