WIP: bindings for LVGL v8
LVGL object constructor can recieve any arguments, not only parent/copy. For that purpose the 'create' function is wrapped and called when constructing the object. All arguments are required for a constructor, so 'parent' must be provided even if it's 'None'

bugfix: create types before functions

Started fixing AdvancedDemo.py, but still work in progress