Add in support for gc9a01 displays
I have added a driver for gc9a01 displays as a child of ili9XXX as it works in a very similar manner.

Two caveats, I have not been able to delve in to the c code and so hybrid mode is not currently supported. Also my display does not seem to match the datasheet for write direction and colour mode (both are reversed) and so I have written it this way in the driver but it may need to be updated if my display is an odd case