Add subscript operator to structs
This allows treating structs as pointers to structs by adding subscript (square brackets after the struct variable, with an index). A subscript would return the array element at the specified position, in the same way C pointers behave. In that sense, the following are always true: s[0] == s, s[0] != s[1]

When creating a new struct, if the constructor first argument is an int - it specifies the number of elements in the array. If it's not an int, we expect the first argument to be a struct of the same type to be copied into the newly created struct. If an int argument is missing, then the default array size is 1 (ie. this is not an array, just a struct instance)