gen/gen_mpy.py: bugfix
When passing a bytes/bytearray/memoryview argument to a C function that expects a pointer, don't copy array contents element by element. Instead, pass the pointer to the data of that bytes/bytearray/memoryview. This improves performance and also allows passing bytes/bytearray/memoryview *by pointer* such that the called function is could modify the data and that would be reflected back on the bytes/bytearray/memoryview that was passed