Support opaque structs
For example lv_disp_t.
Add opque structs to "structs" dict but set delcs to [] instead of None, so they will be generated as stucts with 0 fields.
This is needed because even if struct has 0 fields it may have member functions (getters setters for example) so we need attr implemented and call_parent_methods calls in place.