All finished up. Tests now run.
There is a readme in the tests folder where I have outlined what needs to be done to add tests.
It is actually a pretty simple process.

To run the tests you need to be in the root of the micropython directory. This is the command you will run

```
python3 lib/lv_bindings/tests/__init__.py --artifact-path=lib/lv_bindings/tests/artifacts --mpy-path=ports/unix/build-standard/micropython
```

the paths have to be relative to where ever you are running the tests from. so you can change the paths if wanted. the artifacts folder must exist prior to running the tests.